scripts/requirements.txt
openai
mysql.connector
random